+++ src/lts/__init__.py
@@ -1,0 +1,7 @@
+"""
+云日志服务(LTS)模块
+"""
+
+from .client import LTSClient
+
+__all__ = ['LTSClient']