+++ src/lts/commands.py
@@ -1,0 +1,11 @@
+"""
+云日志服务(LTS)命令行接口
+"""
+
+import click
+
+
+@click.group()
+def lts():
+    """云日志服务(LTS)管理"""
+    pass