+++ src/oceanfs/__init__.py
@@ -1,0 +1,5 @@
+"""海量文件服务(OceanFS)客户端"""
+
+from .client import OceanFSClient
+
+__all__ = ['OceanFSClient']