+++ src/oceanfs/client.py
@@ -1,0 +1,22 @@
+"""海量文件服务(OceanFS)客户端"""
+
+from typing import Dict, Any, Optional
+from core import CTYUNClient
+from auth.eop_signature import CTYUNEOPAuth
+from utils import logger
+
+
+class OceanFSClient:
+    """天翼云海量文件服务(OceanFS)客户端"""
+
+    def __init__(self, client: CTYUNClient):
+        """
+        初始化海量文件服务客户端
+
+        Args:
+            client: 天翼云API客户端
+        """
+        self.client = client
+        self.service = 'oceanfs'
+        self.base_endpoint = 'oceanfs-global.ctapi.ctyun.cn'
+        self.eop_auth = CTYUNEOPAuth(client.access_key, client.secret_key)