+++ src/oceanfs/commands.py
@@ -1,0 +1,9 @@
+"""海量文件服务(OceanFS)命令行接口"""
+
+import click
+
+
+@click.group()
+def oceanfs():
+    """海量文件服务(OceanFS)管理"""
+    pass