+++ src/sfs/__init__.py
@@ -1,0 +1,5 @@
+"""弹性文件服务(SFS)客户端"""
+
+from .client import SFSClient
+
+__all__ = ['SFSClient']