+++ src/sfs/commands.py
@@ -1,0 +1,9 @@
+"""弹性文件服务(SFS)命令行接口"""
+
+import click
+
+
+@click.group()
+def sfs():
+    """弹性文件服务(SFS)管理"""
+    pass